% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_CI.R
\name{average_CI}
\alias{average_CI}
\title{Confidence Interval Averages}
\usage{
average_CI(inputfile)
}
\arguments{
\item{inputfile}{Path to the input file}
}
\value{
A list of 20 average confidence intervals based upon the trained subsample
}
\description{
Finds 20 sets of average confidence intervals based upon the trained subsample of data.
}
