% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmodel.R
\name{buildmodel}
\alias{buildmodel}
\title{Build Logistic Regression Model}
\usage{
buildmodel(inputfile)
}
\arguments{
\item{inputfile}{Path to the input file}
}
\value{
A logistic regression model (glm) of the input file
}
\description{
This function loads a clean dataset as a matrix. It generates a set seed to randomly sample the data (without replacement)
to split said data into a train and prediction subsample. It builds a logistic regression model based upon the trained subsample.
}
