% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_interval.R
\name{pred_CI}
\alias{pred_CI}
\title{Build Predicted Intervals}
\usage{
pred_CI(inputfile)
}
\arguments{
\item{inputfile}{Path to the input file}
}
\value{
A logistic regression model (glm) of the input file
}
\description{
This builds a predicted interval based on the glm of our trained dataset.
}
